Replace ParsedHttpSource with HttpSource
Filters migrated to a new file `DoujinHentaiFilters.kt`
